Removing the arm64 download for now
There has been a handful of reports that this isn't able to download
pre-arm64 versions (<2.14.1). I tried to add a version parser that would
switch back to x64 if the version is lower, but according to users that
didn't seem to work either. I don't know the benefits of downloading the
arm64 version at this time so I'm going to revert the change.